Fix code review issues and repair failing tests

+++ android/app/src/main/kotlin/com/rainguard/MainActivity.kt
@@ -130,7 +130,7 @@ class MainActivity : FlutterActivity() {
                 }
                 "getBubblePosition" -> {
                     val (x, y) = bubbleManager.getPosition()
-                    result.success(hashMapOf("x" to x, "y" to y))
+                    result.success(hashMapOf("x" to x.toDouble(), "y" to y.toDouble()))
                 }
                 "isBubbleVisible" -> {
                     result.success(bubbleManager.isShowing())
